pmove: pass only origin into NudgePosition, instead of pmove
Slightly modified the behaviour of this function. Previously it would
always round to the nearest 1/8th coordinate towards zero, whereas now
I round to the nearest coordinate. Also, there would previously always
be at least one nudge, but I test the position once before starting
that process.

Hopefully, haven't broken any weird behaviour that depends on
this. Got to be careful when dealing with the physics of Quake!

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>