pmove: pull out	the list(stack) of physents from playermove_t
The information	about physents is kind of separate from	the rest of
the player move information - it is also set up before doing
movement/prediction and should be read-only during those calculations.

Most places now	    get a pointer to the "pestack" (kind of a stack due to
adding player entities to the top and popping them off again) passed
in rather than referring to the global data directly.

Still kind of a mess, but should be able to clean up further from
here.

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>