pmove: use more appropriate types for pmove/pstate onground member
Make the playermove onground member a pointer to the physent - the
server wants to know which entity (edict) the player is standing
on. Make the onground member of the player_state_t a boolean, since
nobody cares what you're standing on, only if you are air/water bourne
or not (view.c).

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>