build: include dependency files from arch build dirs
Need to include the dep files from the standard build dirs and the arch
build dirs, so two calls to wildcard here. Using := to assign the value
also gives us a nice speedup to recompiles (reading lots of small files
like this is very slow on Msys/MinGW on Windows).

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>