build: fix ghetto unix2dos conversion to work on msys
Pull in the same fix used on TyrUtils.

The only thing that seems to be able to convert \n to \r\n is awk, so use
that everywhere.

I don't understand why, but echoing the verbose command on OSX echos a
literal \r and messes up the output. Adding an extra \ escape fixes it
there, but then Linux/Msys echo the extra \... I think the OSX behaviour
might be broken here, oh well...

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>