vid: use window centering code from wgl for vid_win driver
For now we ignore the vid_win_x/y cvars and center the window on load.
This is probably going to be what people want in 99% of cases, but I can
imagine it being useful to be able to script the window position on the
command line for some obscure cases, so I will restore that functionality
later.

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>