draw: don't scribble pic data with glpic_t
Switch things around a bit so that we are not scribbling over the first
few bytes of pic data when we store the glpic_t. Change the definition of
pic_t so that data is a pointer and embed a pic_t in the glpic_t. Then we
just need to place the glpic_t on the hunk, copy the pic->width and
pic->height and point data at the actual pic data provided by
W_GetLumpName.

Unfortunately this made things slightly more complicated on the software
renderer side. Maybe we should be modifying the pic wad at load time
instead?

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>