model: factor out (and fix) bsp submodel setup
Broke this in a previous patch where I forgot to move the submodel
pointer on to the next model.  This whole function was done in a very
confusing way originally.  Make it clear that we are just copying some
fields from the world model and fixing up the remaining fields from
the submodel dmodels.

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>