zone: add a new API Hunk_AllocExtend
Allow the low hunk allocation to be extended.  And require the base
pointer for the hunk to be passed in so we can test that matches what
the hunkstate remembers as the bottom hunk.

Intended to be used to group lots of small allocations together in a
single hunk (thinking of glpoly_t as the first instance).

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>