model: move mod_known onto the hunk, split model cache
Well, the only caching that is actually done is for the alias models.
Sprites and BSP models get forgotten as soon as the next map loads.
So allocate some space (currently hard coded to 512) for alias model
caching below host_hunkelevel and put bsp and sprite model structs
onto the hunk with the rest of the level data.

This structure still needs some work, but at least we should no longer
run out of space for models as long has we have some space on the
hunk.

The alias model overflow thing is probably a bit of overkill or the
wrong way to handle it - we should be doing some kind of LRU and
throwing away things we no longer need as well.

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>