demo: use sscanf to parse cd track number
My previous patch to CL_PlayDemo_f was a bit to restrictive in
checking the cd track number and broke hipdemo1 from the official
expansion pack due to this demo having a leading space before the
track number (recorded with a different build of the engine, I guess).

Instead of parsing the track number manually, we pull out up to 12
characters looking for a newline, then use sscanf to parse the track
number.  If we didn't find a newline or the track number doesn't parse
as an integer, we assume the demo file is invalid.

Hopefully this is strict enough to filter our files that really are
corrupt but not so strict as to break otherwise working demos.

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>