common: make com_token a const pointer
Make the buffer for com_token private to common.c and only allow other
files to access it via a const pointer.  Add checks for overflowing
the token buffer; in case this happens we just truncate the token, but
continue to parse from the source until completing the source token.

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>