model: remove model size limits from alias_model.c
Because we flatten out the skin/frame groups in the final in-memory
structures, we need some temporary buffers to simplify loading.  These
were just static buffers but now dynamically allocated on the hunk.

Unfortunately to work out how much space we need, we need to walk over
the whole model structure first, but I suppose it's not too bad.

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>