model: save alias skin pixel data into the cache
To simplify things a bit, we will save the 8-bit texture data for all
alias models in the cache.  Wastes a little bit of memory, but we can
remove some hacks for saving the player skins and doing color
translations. Also, this will simplify reloading gl textures when the
context has to be re-created for vid mode changes.

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>