textures: check for npot extension and use if available
Created the gl_extensions.c file to pull some of the common stuff into
where appropriate.  Non-power of two textures will be allowed by default,
with the option to go back to old behaviour by setting command line option
"-nonpot" or cvar 'gl_npot 0'.  The cvar is a bit of a hack, as existing
textures won't be reloaded at this stage.

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>