model: support two types of BSP2 models
Turns out there are two types of "BSP2" format - one that I think was
created by MH and implemented in the RMQ project (this is the one I
implemented previously) and one by LordHavoc which exteneded the
coordinate limits beyond 32k, implemented in Darkplaces.  The
Darkplaces version of BSP2 will get the name BSP2 and the RMQ version,
which is more limited, will get the name BSP2rmq.

Fairly naive implementation which duplicates a bunch of code.  Can be
refactored and tidied up more later.

Signed-off-by: Kevin Shanahan <kmshanah@disenchant.net>