Always try and allocate the first available free block in the zone
Previous behaviour was always to start searching for free blocks after
the last successful allocation.  This leads to crazy fragmentation as we
will always allocate from the biggest chunk of memory until we have
fragmented the whole zone.

Instead we accept a bit of a hit on allocations as we will be scanning
over more blocks to always allocate as close as possible to the bottom
of the zone.  But we then ensure we maintain the largest chunk of free
memory possible at the top of the zone.

Signed-off-by: Kevin Shanahan <kevin@shanahan.id.au>