Fix netgraph statistics for packets not-yet-received
We were previously showing all packets sent but not yet responded to
as packet loss, even if no subsequent packets have been received.
This is misleading as we may have just sent the packet out and the
response will still arrive in a timely manner.

Instead, we will show packets not-yet-received with their current
latency.  If we receive a response to a later packet, then we will
show the previous not-yet-received packets as dropped at that stage.

Signed-off-by: Kevin Shanahan <kevin@shanahan.id.au>