gl: Fix reloading of hud pics
Previous method of calling Sbar_InitPics again appeared to work, but
it was allocating new glpic structs on the hunk above the host
hunklevel.  Besides being wasteful, these were being cleared on level
change, and the status bar became corrupted.

Instead, be more careful about the initialisation of the scrap
textures and keep track of any glpics allocated separately so we can
reload them using the data already on the hunk.