screen: Implement scalable HUD elements
Big batch of changes that allows a separate scale to be used for the
HUD elements (status bar, notification text, console text) than just
the pixel resolution of the main buffer.

In software mode this is fairly crude scaling, although it looks
surprisingly good in most cases.  In OpenGL we get the selected
texture filtering mode applied to the adjusted coordinate system,
which looks quite nice.

- New console command 'hudscale' and variable 'scr_hudscale'
- Fixed 'crosshair 2' mode and made the crosshair correctly scale.
- The -conwidth and -conheight command line parameters (which were
  only available with the wgl renderer anyway) are now removed.