gl: fix status bar redraw issue
When rendering the status bar, the code tries to be optimal by
checking if what was in the old front buffer is still up to date and
skipping the draw in that case.  However, GL implementations are free
to give us triple (or more!) buffers or even never maintain the old
contents after xxxSwapBuffers().  Specifically, the contents of the
back buffer are 'undefined' after a call to glXSwapBuffers.

In the vid struct, vid.numpages is supposed to represent the number of
draw buffers we are rotating through.  Although, commonly
implementations *will* return the old front buffer contents after 1 or
2 swaps, this is not guaranteed, so we may as well just redraw every
frame.  It's not like the status bar is the expensive part of the
rendering anyway.

We may later make some optimisations like rendering the status bar to
a texture so it's just one quad to redraw if nothing changed.