gl: first pass at fullbright pixels support
Support fullbright pixels in GL by creating an overlay texture with
only the fullbright pixels opaque and blending over the top of the
already shadowed base texture.  Incomplete as it only does this for
the bsp models currently and only when multitexturing is enabled.

Credit to FitzQuake for the inspiration to use alternate palettes to
generate the mask textures.  This enabled a bunch of cleanups in the
texture uploading code by embedding the info about which textures are
alpha in the texture type and palette.