gl: refactor model loading to store lightmap blocks with the brushmodel
Fairly major refactoring that still has a few rough edges, but this
now gets us lightmap blocks stored along with the brushmodel and
allocated dynamically to remove the static limit on the number of
blocks available.

Lack of sharing between models does mean with allocate a whole block
for e.g. ammo box lightmaps, which is kind of crazy.  But we can get
to that later.