gl: add brushmodel surfaces to their material chains in preparation for drawing
Add helper functions for making sure we handle animating textures
correctly with the material chains.  Because we were careful to store
the zero frame of the animation as the material number on the surface,
we can just swap the zero frame chain for the chain of the currently
displaying animation frame before iterating over the brushmodel
surfaces to gather the surfaces into the correct chain.