gl: switch over to material chain drawing
Big code drop here with most of the payoff from all the work lining up
the materials for the BSP models.  We now push everything through a
triangle buffer taking advantage of glDrawElements.  This reduces the
number of draw calls dramatically and can be as much as a 2x speedup.

For submodels (brushmodel entities) we now batch all static entities
into the same draw calls as the world.  We have to use separate draw
calls for any moving models due to the change in translation matrix.

Definitely still a few broken things to fix, but a big step forward.