osx: bring back ability to build on osx 10.5
We can now build on 10.5 again, which means we can have universal
binaries for intel/ppc.  These actually work on modern OSX as well,
but apparently everything except x64 is now dead according to Apple.

This currently relies on an old version of SDL2, since libsdl dropped
support for 10.5 in more recent versions as well.  Perhaps I should do
a native cocoa layer in the future, however this will do for now.