client: fix issue with duplicate entries in visedicts
The code that populates the visible edicts from entity fragments
(efrags) was relying on a change in r_framecount implying that the
visible entities list needed to be re-populated.  However that was not
true in certain corner cases, such as restarting a level after player
death.  In GLQuake, rendering the loading plaque means an extra world
frame is rendered without resetting the visible edicts (done in
CL_ReadFromServer()), so duplicate entries were present on the list.

This previously would have harmlessly gone by with some entities being
rendered twice (with no real visible effect, maybe if there was
transparency involved) however the material chain rendering can fail
badly if this happens, as it can create loops in the material chain.

Instead, tie the population of visible edicts directly to when the
list was reset by using a dedicated frame counter for that.