gl: remove strange logic in R_RecursiveWorldNode
As far as I can see this is bogus, as if it ever has any effect it
will be to make us recurse down the same side of the BSP twice in a
row instead of visiting both child nodes.  Odd.  I don't know if I
wrote that, but it goes back beyond git history and even my manual
patches before that...