gl: fix corruption of QW skin textures during upload
Because the skin pixels were being loaded via the data cache, when the
texture upload process allocated memory on the hunk to do resampling
it was possible the skin pixels were 'evicted' from the cache and the
texture resampling would be trying to use the same memory to store
it's results.

Copy the skin pixels to the hunk before processing the texture to work
around this.