sw: render mask/fence textures in second pass of edge renderer
After generating the surface and edge list for the scene, we save a
copy for a potential 2nd pass.

During the first span generation pass, we skip over any surfaces with
the SURF_DRAWFENCE flag and return a flag for each scanline that
contains such surfaces.  If any scanlines with SURF_DRAWFENCE were
found, restore the unmodified surf/edge lists and re-generate spans
for just the flagged surfaces.

And finally, we add a simple span drawing function that handles the
mask/fence textures properly by skipping writes to the z/color buffer
when masked pixels are crossed.

Heavily inspired by the approach used in the Hexen II software
renderer for handling transparency.