sw: get edge_t back down to 32 bytes
I think when I changed this long ago I didn't understand that the
surfs being referred to in the edge_t struct are just the edge
scanning surfaces and not the surface numbers in the BSP.

So, we don't really need more than 16 bits unless we plan to have >65k
surfaces in the view at the same time and still expect the software
edge renderer to keep up!  Maybe it could, but for now it's nicer fit
back into cache lines again.