sw: Update assembly routines to skip fence surfs during span generation
Juggle a few registers to make room for us to accumulate a return
value containing the flagged surface values.  I'm sure I ruined some
of the original version's artful use of the original Pentium CPU
cache/pipeline/whatever, but it shouldn't be too much of a loss.
Certainly not noticable on systems I have currently...