sw: improve memory usage of bmodel clipping
We want the use of the edges/verts be kept when going down the
recursive clip process, but freed back up when we come up to continue
clipping on the other side of the node.  This make for more than 2x
improvement in usage of the stack resources for some test cases.

Reduce the default stack allocation again, since I can't trigger
anything that will overflow it now.  But still should be made dynamic
at some stage.