alpha: add depth sorting primitives for rendering translucent entities/surfaces
Add a new 'depthchain' primitive which is just a depth-sorted list of
all the surfaces/entities that need to be drawn in back-to-front order
for correct alpha blending.

Should give good results in most cases, but of course intersecting
translucent surfaces won't be quite right unless we clip all the
translucent polys against each other and re-sort.

We don't actually draw the translucent entities yet, so things will
look completely broken for now...