alpha: improve depth sorting keys
For surfaces, we need to know the point at which the surface is
deepest into the scene to propertly depth sort.  For models, it's a
little more accurate to use the centre of their bounding box as their
depth point, since the origin is not necessarily near the centre of
their visible region.