gl: use fitzquake style water warp rendering technique
Similar technique to FitzQuake, render a warped version of the liquid
texture to the back-buffer and copy into a texture that holds the
pre-warped image.  Then we don't need to subdivide water surfaces
anymore.  So we pay a bit more to pre-warp the image, but for large
expanses of water the rendering is much more efficient (and better
quality!)