sys: Add Sys_MakeCodeUnwritable() to restore code write protection after patching
Rather than leaving some of the assembly paths writable for the whole
execution of the program, restore those write protections right after.
This will reduce the possibility of an exploit elsewhere in the code
being able to insert malicious code into the writable assembly
regions.

There may be some performance impact on older hardware, but it is not
noticable on my current systems.  Who knows if it's even worth
maintaining these assembly code paths for much longer anyway.