gl: add gl_max_textures cvar to specify the size of the texture cache
Can be resized, but requires a vid_restart to take effect.  Make this
a CVAR_VID flagged variable so that it is loaded in the early vid
config.  Had to make sure the cvar was registered early too.  This is
a bit hacky for now, but we should just change the initialisation
order of all cvars/commands to be earlier than other subsytem init.