net: make game discovery via network broadcast more reliable
When multiple IP addresses are available on the machine, sending
broadcasts via the generic broadcast address is unreliable.  The
packet could go anywhere.

Instead we now attempt to enumerate all network addresses on the
machine and if we bind to INADDR_ANY, then we broadcast across all
(local) networks.

Additionally, we are now careful that the CCREP_SERVER_INFO responses
contain the ip address the corresponds to the network on which we are
responding.