host: rework filesystem init so we can try several basedir candidates
To make it as easy as possible to e.g. use your Steam installation of
Quake, I wanted to have the executable look for a list of potential
candidate directories where it could find the game data files.

So on Windows only for now, it will look for the Steam Quake
installation folder via the registry and a few other 'likely'
candidate locations where quake could be installed.