common: use a more useful buffer size from COM_StackLoadFile callers
There are practically NO files that will load into a 1k stack buffer,
which is what we were using before.  A 64k stack buffer fits a lot of
short sound files and simple models, which actually helps to reduce
the amount of pushing around of cache items while loading assets.