gl: fix for warp texture target when the target is in the texture cache
Without this, when we were hitting the cache with a warp target
texture the target size was returned as the original texture size, not
the upsampled target texture.

This is just a symptom of a bad design for the texture cache api in
general... to be overhauled at some stage.