sound: add a smooth fade out of sound effects cut off by replacement
Running with backing buffer at a higher sample rate, I found it
occasionally noticable that when a sound in a channel is replaced by
another sound mid-playback there was an audible pop or click.

Instead of hard killing the playing sound, use a smooth but rapid fade
out curve on the current channel and select a new channel for the
replacement sound.