sound: keep dma buffer locked for the whole sound paint process
Because we just check the dma position at the start of the sound paint
process, we should keep the dma buffer locked until we have finished
mixing and painted the sound out to the buffer.

If this keeps the buffer locked too long, we would have to look at
separately locking specific regions of the buffer.