sw: allocate extra edges/surfs on the high hunk
Now that we are allocating overflow for alias models and sound effects
on the low hunk, we can't just reset the hunk level whenever we need
to allocate more surfaces or edges for the software renderer.
Rearrange things slightly and allocate them on the high hunk instead,
along with the other video buffer data.