gl: Implement simple vertex program for lerping alias models on the GPU
Using the ancient GL_ARB_VERTEX_PROGRAM extension with text programs
written in assembly because I want to support some older machines with
only OpenGL 1.3(ish) available.

This also tidies up some of the vertex buffer handling so shared
buffers are initialised globally and the per-model buffers are
initialised as models are loaded (and freed when flushed from cache).