gl: dynamically allocate vertex buffers for BSP rendering
Rather than the stack buffer I was previously using, allocate vertex
buffers on the hunk.  This means we can have enough space for very
large batches if needed.

This also introduces some basic state tracking for the vertex buffer
which is intended to help when the vbo rendering paths is implemented.