gl: add support for compressed textures
This helps a lot on systems with low amounts of GPU memory.  There is
a bit of a quality hit for certain textures, but for the most part it
is not too noticable.  Enabled by default for now, but would be nice
to make this controlled via console variable instead of command line
only.