console: draw conback string using alt text instead of palette shift
The original code just shifted the palette index up on the conback
text, but this only works for the original charset and can have very
weird effects when a custom charset is used.  Instead, just use the
alt text which is close to the same colors on the standard charset and
will at least look sensible on custom charsets.