misc: fov and screen aspect ratio fixes, translucent view weapon
Misc fixes for FOV handling and setting the correct aspect ratio of
the screen.  Removed the old hacks that were trying to get it right
for 320x200 on old 4:3 monitors, etc.

Make the view weapon draw semi-transparently when invisibility powerup
is active, or if the user has configured r_drawviewmodel with a value
less than one.  Nice addition from FQ/QS.