gl: fix sbar glpics embedded in scrap texture after reload
Instead of storing the gl texture id with each glpic embedded in the
scrap texture, the glpic has the scrap information embedded and that
information is used when binding the texture.  This gets rid of the
situation where previously the texture id on each glpic had to be
updated when texutures were reloaded.  Now only the texture ids on the
scrap textures need to be updated.