sw: implement scaling of the rendering resolution relative to window size
Instead of always insisting on rendering the window at full
resolution, allow rendering and half/quarter or other resolutions and
then upscaling to the window size.  Rendering at lower resolution and
scaling up to window size is almost always a performance win, but
sometimes might also be wanted for aesthetic reasons.

This patch implements the menu interface and the win32 driver support.