screen: restore old behaviour of viewsize 110/120
When adding the transparent sbar, I had changed the meaning viewsize
that remove the sbar entirely from 120 to 130.  However, people do
write quakec code that uses viewsize 120 to remove the sbar for a
cutscene.  So, restore that behaviour and instead viewsize 100 will be
the sbar with no side fill, 90 will be the full sbar with side fill
and then 90 and lower will reduce the main view rect.