gl: fix reloading of warp lookup texture
The warp lookup texture was not being reloaded correctly on
vid_restart.  Noticed that the notexture texture was in the same
situation.  Additionally, need to reset the cached known bound
textures when textures are freed, otherwise we can sometimes not
re-load the charset texture if it was bound just before vid_restart.