sdl: Fix writing outside the allocated SDL texture used for blitting
For odd-sized resolutions, we have a situation where the backing
buffer is rounded up to a multiple of 8 size, but the output buffer is
not.  So, be careful when blitting to not write outside the allocated
surface/texture.

Also, we have no need to round height to multiple of 8 (only width),
so stop doing that!