alias: implement MF_HOLEY rendering for alias models
Handle the MF_HOLEY model flag from MarkV/QSS which indicates models
are to be rendered with color palette index 255 being transparent
(similar to 'fence' textures on BSP models).

Add rendering paths for GLQuake and the software renderer.  This also
fixes some glitches in the rendering of translucent alias models in
the software renderer (errors in the assembly for the subdivided
rendering path).